+++ src/components/page-types/oodInteriorPage.js
@@ -2,79 +2,20 @@ import React from 'react'
 import SbEditable from 'storyblok-react'
 import RichTextField from '../../utilities/richTextField'
 import Components from "../components"
+import HeaderMinimal from '../partials/headerMinimal'
 import HeaderNoImage from '../partials/headerNoImage'
+import HeaderWithImage from '../partials/headerWithImage'
+import BodyLeftSidebar from '../partials/bodyLeftSidebar'
+import BodyNoSidebar from '../partials/bodyNoSidebar'
 import IconCardSection from '../partials/iconCardSection'
+import Footer from '../partials/footer'
 import { Helmet } from 'react-helmet'
-
-const HeaderWithImage = (props) => (
-  <SbEditable content={props.blok}>
-    <header className={`ood-interior-page__header ood-interior-page__header--has-image`}>
-      <div className={`ood-interior-page__header-title-wrapper su-pt-6 su-pb-5 su-bg-${props.blok.headerBackgroundColor}`}>
-        <div className={`centered-container flex-container`}>
-          <h1 className="ood-interior-page__title flex-lg-6-of-12 flex-xl-5-of-12 flex-2xl-6-of-12 su-serif su-text-white su-text-align-left">{props.blok.title}</h1>
-          <figure className="su-media flex-lg-6-of-12 flex-xl-7-of-12 flex-2xl-6-of-12 ood-interior-page__header-media">
-            <div className="su-media__wrapper su-aspect-ratio su-aspect-ratio--3x2">
-              <img className="ood-interior-page__image" src={props.blok.headerImage.filename ? props.blok.headerImage.filename : ""} alt="" />
-            </div>
-          </figure>
-        </div>
-      </div>
-      <div className={`ood-interior-page__header-intro-wrapper su-py-6 su-bg-white`}>
-        <div className={`centered-container flex-container`}>
-          {props.blok.intro && (
-            <div className="su-intro-text ood-interior-page__intro flex-xl-8-of-12">
-              <RichTextField data={props.blok.intro}/>
-            </div>
-          )}
-        </div>
-      </div>
-    </header>
-  </SbEditable>
-);
-
-const HeaderMinimal = (props) => (
-  <SbEditable content={props.blok}>
-    <header className={`ood-interior-page__header ood-interior-page__header--minimal su-bg-fog-light su-py-7`}>
-      <div className={`centered-container`}>
-        <h1 className="ood-interior-page__title su-serif su-text-align-center">{props.blok.title}</h1>
-      </div>
-    </header>
-  </SbEditable>
-);
-
-const BodyLeftSidebar = (props) => (
-  <SbEditable content={props.blok}>
-    <aside className="ood-interior-page__body-sidebar flex-lg-4-of-12 flex-xl-3-of-12">
-      {props.blok.contentMenu && props.blok.contentMenu.map((blok) => React.createElement(Components(blok.component), {
-        key: blok._uid,
-        blok: blok
-      }))}
-    </aside>
-    <div className="ood-interior-page__body-content flex-lg-8-of-12">
-      {props.blok.pageContent && props.blok.pageContent.map((blok) => React.createElement(Components(blok.component), {
-        key: blok._uid,
-        blok: blok
-      }))}
-    </div>
-  </SbEditable>
-);
-
-const BodyNoSidebar = (props) => (
-  <SbEditable content={props.blok}>
-    <div className="ood-interior-page__body-content flex-lg-8-of-12 flex-xl-7-of-12 flex-2xl-6-of-12 su-mx-auto">
-      {props.blok.pageContent && props.blok.pageContent.map((blok) => React.createElement(Components(blok.component), {
-        key: blok._uid,
-        blok: blok
-      }))}
-    </div>
-  </SbEditable>
-)
 
 const OodInteriorPage = (props) => {
   return (
     <>
-      <Helmet><title>{`${props.blok.title} | Giving to Stanford`}</title></Helmet>
       <SbEditable content={props.blok}>
+        <Helmet><title>{`${props.blok.title} | Giving to Stanford`}</title></Helmet>
         {props.blok.localHeader && props.blok.localHeader.map((blok) => React.createElement(Components(blok.component), {
           key: blok._uid,
           blok: blok
@@ -120,16 +61,7 @@ const OodInteriorPage = (props) => {
             </footer>
           </article>
         </main>
-        <footer>
-          {props.blok.localFooter && props.blok.localFooter.map((blok) => React.createElement(Components(blok.component), {
-            key: blok._uid,
-            blok: blok
-          }))}
-          {props.blok.globalFooter && props.blok.globalFooter.map((blok) => React.createElement(Components(blok.component), {
-            key: blok._uid,
-            blok: blok
-          }))}
-        </footer>
+        <Footer {...props}/>
       </SbEditable>
     </>
   )
